Particle System & Toast Notification Fixes and New Wand
Fixed some of the existing particle systems and added particle systems to some of the Mushrooms that still needed them.

Fixed position and size of toast notification displayed when player gets a stat boost.

Added a new wand by copying an existing weapon, but have yet to modify it for in-game use.